Move remaining configuration
Now the dependency on octokit is removed from "Engine" we can move the remaining configuration objects into abstactions
