+++ NuKeeper.Integration.Tests/Engine/RepositoryFilterTests.cs
@@ -3,7 +3,6 @@ using System.Threading.Tasks;
 using NSubstitute;
 using NuKeeper.Abstractions.Configuration;
 using NuKeeper.Abstractions.Logging;
-using NuKeeper.Configuration;
 using NuKeeper.Engine;
 using NuKeeper.GitHub;
 using NUnit.Framework;
