+++ NuKeeper/Commands/LocalNuKeeperCommand.cs
@@ -1,5 +1,4 @@
 using McMaster.Extensions.CommandLineUtils;
-using NuKeeper.Configuration;
 using NuKeeper.Inspection.Logging;
 using System.IO;
 using NuKeeper.Abstractions.Configuration;
