+++ NuKeeper/Commands/OrganisationCommand.cs
@@ -1,6 +1,5 @@
 using McMaster.Extensions.CommandLineUtils;
 using NuKeeper.Abstractions.Configuration;
-using NuKeeper.Configuration;
 using NuKeeper.Engine;
 using NuKeeper.Inspection.Logging;
