+++ NuKeeper/Engine/GitHubRepositoryEngine.cs
@@ -4,7 +4,6 @@ using LibGit2Sharp;
 using NuKeeper.Abstractions.Configuration;
 using NuKeeper.Abstractions.DTOs;
 using NuKeeper.Abstractions.Logging;
-using NuKeeper.Configuration;
 using NuKeeper.Git;
 using NuKeeper.Inspection.Files;
 using NuKeeper.Inspection.Logging;
