+++ NuKeeper/Engine/IForkFinder.cs
@@ -1,7 +1,6 @@
 using System.Threading.Tasks;
 using NuKeeper.Abstractions.Configuration;
 using NuKeeper.Abstractions.DTOs;
-using NuKeeper.Configuration;
 
 namespace NuKeeper.Engine
 {
