+++ NuKeeper/Engine/IGitHubRepositoryDiscovery.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using NuKeeper.Configuration;
+using NuKeeper.Abstractions.Configuration;
 using NuKeeper.GitHub;
 
 namespace NuKeeper.Engine
