+++ NuKeeper/Engine/Packages/IPackageUpdateSelection.cs
@@ -1,8 +1,8 @@
 using NuKeeper.Inspection.RepositoryInspection;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using NuKeeper.Abstractions.Configuration;
 using NuKeeper.Abstractions.DTOs;
-using NuKeeper.Configuration;
 using NuKeeper.Update.Selection;
 
 namespace NuKeeper.Engine.Packages
