+++ NuKeeper/Engine/Packages/PackageUpdater.cs
@@ -1,7 +1,6 @@
 using NuKeeper.Abstractions.DTOs;
 using NuKeeper.Abstractions.Logging;
 using NuKeeper.Abstractions.NuGet;
-using NuKeeper.Configuration;
 using NuKeeper.Git;
 using NuKeeper.GitHub;
 using NuKeeper.Inspection.RepositoryInspection;
@@ -9,6 +8,7 @@ using NuKeeper.Update;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using NuKeeper.Abstractions.Configuration;
 
 namespace NuKeeper.Engine.Packages
 {
